feat: url에 따른 키워드 추천이 되도록 기능을 분화시켜 커넥터로 작성
