+++ src/ConnectorGlobal.ts
@@ -32,8 +32,9 @@ export namespace ConnectorGlobal {
     GOOGLE_TEST_SECRET: string;
 
     // GOOGLE ADS
-    GOOGLE_ADS_DEVELOPER_TOKEN: string;
     GOOGLE_ADS_ACCOUNT_ID: string;
+    GOOGLE_ADS_DEVELOPER_TOKEN: string;
+    GOOGLE_ADS_PARENT_SECRET: string;
 
     // NAVER
     NAVER_CLIENT_ID: string;
