+++ src/api/functional/connector/google_ads/index.ts
@@ -4,78 +4,4 @@
  * @nestia Generated by Nestia - https://github.com/samchon/nestia
  */
 //================================================================
-import type { IConnection, Primitive, Resolved } from "@nestia/fetcher";
-import { NestiaSimulator } from "@nestia/fetcher/lib/NestiaSimulator";
-import { PlainFetcher } from "@nestia/fetcher/lib/PlainFetcher";
-import typia from "typia";
-
-import type { IGoogleAds } from "../../../structures/connector/google_ads/IGoogleAds";
-
-/**
- * @controller GoogleAdsController.generateKeywordIdeas
- * @path POST /connector/google-ads/generateKeywordIdeas
- * @nestia Generated by Nestia - https://github.com/samchon/nestia
- */
-export async function generateKeywordIdeas(
-  connection: IConnection,
-  input: generateKeywordIdeas.Input,
-): Promise<generateKeywordIdeas.Output> {
-  return !!connection.simulate
-    ? generateKeywordIdeas.simulate(connection, input)
-    : PlainFetcher.fetch(
-        {
-          ...connection,
-          headers: {
-            ...connection.headers,
-            "Content-Type": "application/json",
-          },
-        },
-        {
-          ...generateKeywordIdeas.METADATA,
-          template: generateKeywordIdeas.METADATA.path,
-          path: generateKeywordIdeas.path(),
-        },
-        input,
-      );
-}
-export namespace generateKeywordIdeas {
-  export type Input = Primitive<IGoogleAds.IGenerateKeywordIdeaInput>;
-  export type Output = Primitive<IGoogleAds.IGenerateKeywordIdeaOutput>;
-
-  export const METADATA = {
-    method: "POST",
-    path: "/connector/google-ads/generateKeywordIdeas",
-    request: {
-      type: "application/json",
-      encrypted: false,
-    },
-    response: {
-      type: "application/json",
-      encrypted: false,
-    },
-    status: null,
-  } as const;
-
-  export const path = () => "/connector/google-ads/generateKeywordIdeas";
-  export const random = (
-    g?: Partial<typia.IRandomGenerator>,
-  ): Resolved<Primitive<IGoogleAds.IGenerateKeywordIdeaOutput>> =>
-    typia.random<Primitive<IGoogleAds.IGenerateKeywordIdeaOutput>>(g);
-  export const simulate = (
-    connection: IConnection,
-    input: generateKeywordIdeas.Input,
-  ): Output => {
-    const assert = NestiaSimulator.assert({
-      method: METADATA.method,
-      host: connection.host,
-      path: path(),
-      contentType: "application/json",
-    });
-    assert.body(() => typia.assert(input));
-    return random(
-      "object" === typeof connection.simulate && null !== connection.simulate
-        ? connection.simulate
-        : undefined,
-    );
-  };
-}
+export * as generateKeywordIdeas from "./generateKeywordIdeas";
